deps: update V8 to 8.1.307.20
PR-URL: https://github.com/nodejs/node/pull/32116

+++ deps/v8/src/debug/debug-property-iterator.h
@@ -46,9 +46,9 @@ class DebugPropertyIterator final : public debug::PropertyIterator {
   enum Stage { kExoticIndices = 0, kEnumerableStrings = 1, kAllProperties = 2 };
   Stage stage_ = kExoticIndices;
 
-  uint32_t current_key_index_ = 0;
+  size_t current_key_index_ = 0;
   Handle<FixedArray> keys_;
-  uint32_t exotic_length_ = 0;
+  size_t exotic_length_ = 0;
 
   bool calculated_native_accessor_flags_ = false;
   int native_accessor_flags_ = 0;
